Fix infinite loop !!! after a long long debug
A Path object have different meta_info during its life. Early in its
life, it has very basic meta_info, based on the filename (and perhaps
on metadata files). Then when the SourceHandler is known for the
file, the default metadata for this SourceHandler are added. And then
when the file is read, the JSON header is used as the definite source
for meta info.

There is another mechanism in webgen that loop the process until no
more node meta_info were changed. To detect in a meta_info has changed
(SourceHandler.meta_info_changed?), webgen compares the meta_info found
on a path in its early stage (before the SourceHandler is known) with
the cache (that has meta_info at the same advancement stage)

Note: the key for the cache is the path string and the SourceHandler
class name.

The infinite loop was caused by the Index SourceHandler that created
Feed nodes. The problem is that Index.create_node was given a path that
has already been populated with the default meta_info for Index
SourceHandler, and index further modified the meta_info with the
content of the file. When the feed nodes were created, they were given
a path that had too much information in its meta_info Hash. And this
wrong meta_info is put in the cache.

Later on, in meta_info_changed?, the more complete data from the cache
is compared with the basic data created from the basic info for the
path, and it detects the meta_info has changed, and a new run is
needed.